backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
supabase==2.13.0
python-dotenv==1.0.1
ezdxf==1.3.4
python-docx==1.1.2
anthropic==0.40.0
pyproj==3.7.1
shapely==2.0.6
pyshp==2.3.1
python-multipart==0.0.19
httpx==0.28.1
defusedxml==0.7.1
slowapi==0.1.9
